Move concat and subsetting operations into groupby operation
This seems to be beneficial as dask does not need to keep track of the
whole array anymore in order to allow concat in a later step. It is
a bit slower though as we need more concat operations in total (first
for each group, then to fuse the groups into partitions).